Rename NetworkService -> BasicNetworkService & NetworkServiceProviding -> NetworkService
