+++ Source/ModifyRequestNetworkService.swift
@@ -31,17 +31,17 @@ import Dispatch
 
 /// `ModifyRequestNetworkService` can be composed with a networkService to modify all outgoing requests.
 /// One could add auth tokens or API keys for specifics URLs.
-public final class ModifyRequestNetworkService: NetworkServiceProviding {
+public final class ModifyRequestNetworkService: NetworkService {
     
     private let requestModifications: Array<(URLRequestConvertible) -> URLRequestConvertible>
-    private let networkService: NetworkServiceProviding
+    private let networkService: NetworkService
     
     /// Creates an insatcne of `ModifyRequestNetworkService`.
     ///
     /// - Parameters:
     ///   - networkService: a networkservice.
     ///   - requestModifications: array of modifications to modify requests.
-    public init(networkService: NetworkServiceProviding, requestModifications: Array<(URLRequestConvertible) -> URLRequestConvertible>) {
+    public init(networkService: NetworkService, requestModifications: Array<(URLRequestConvertible) -> URLRequestConvertible>) {
         self.networkService = networkService
         self.requestModifications = requestModifications
     }
