+++ Tests/DBNetworkStackTests/NetworkServiceTest.swift
@@ -43,7 +43,7 @@ class NetworkServiceTest: XCTestCase {
         ]
     }()
     
-    var networkService: NetworkServiceProviding!
+    var networkService: NetworkService!
     
     var networkAccess = NetworkAccessMock()
     
@@ -56,7 +56,7 @@ class NetworkServiceTest: XCTestCase {
     }
     
     override func setUp() {
-        networkService = NetworkService(networkAccess: networkAccess)
+        networkService = BasicNetworkService(networkAccess: networkAccess)
     }
     
     func testRequest_withValidResponse() {
